Redesign of score parameters

+++ src/main/org/audiveris/omr/util/DoubleValue.java
@@ -55,7 +55,7 @@ public class DoubleValue
      */
     public DoubleValue (Double value)
     {
-        this.value = value.doubleValue();
+        this.value = value;
     }
 
     /**
